completed route handlers
