+++ src/api/routes/handlers/searchPage.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = (req, res) => {
+  res.render('pages/searches/new');
+};
